Split up tsm_worker.js
We now got the tsplib.js which holds our description of the
TravellingSalesmanSolver and our Network node tree we are working on.

Also the Benchmark got outsourced of the temporary worker file.
Tests got deleted and were added to its own folder and made compatible
with nodeunit. (Therefore nodeunit was added to the dev-dependencies)